feat: Add Microsoft.AspNetCore.Authentication.JwtBearer package
This commit adds the Microsoft.AspNetCore.Authentication.JwtBearer package to the AuctionService project. The package is included as a reference in the AuctionService.csproj file. This package provides JWT bearer authentication support for the AuctionService, allowing it to authenticate requests using JSON Web Tokens. This is necessary for implementing secure authentication and authorization in the AuctionService.

+++ backend/src/Auction/AuctionService/Controllers/AuctionsController.cs
@@ -5,6 +5,7 @@ using AutoMapper;
 using AutoMapper.QueryableExtensions;
 using Contracts;
 using MassTransit;
+using Microsoft.AspNetCore.Authorization;
 using Microsoft.AspNetCore.Mvc;
 using Microsoft.EntityFrameworkCore;
 
@@ -52,12 +53,13 @@ public class AuctionsController : ControllerBase
     }
 
     // CreateAuction
+    [Authorize]
     [HttpPost(Name = "CreateAuction")]
     public async Task<ActionResult<AuctionDto>> CreateAuction(CreateAuctionDto createAuctionDto)
     {
         var auction = _mapper.Map<Auction>(createAuctionDto);
         // placeholder for Seller
-        auction.Seller = "test";
+        auction.Seller = User.Identity.Name;
 
         // Using the Outbox pattern, MessageBroker messages are saved in the Outbox table
         // and are only sent to the MessageBroker after the DB transaction is committed
@@ -74,6 +76,7 @@ public class AuctionsController : ControllerBase
     }
 
     // UpdateAuction
+    [Authorize]
     [HttpPut("{id}", Name = "UpdateAuction")]
     public async Task<ActionResult> UpdateAuction(Guid id, UpdateAuctionDto updateAuctionDto)
     {
@@ -82,7 +85,7 @@ public class AuctionsController : ControllerBase
         .FirstOrDefaultAsync(x => x.Id == id);
         if (auction == null) return NotFound();
 
-        // TODO: Check Seller == Username
+        if (auction.Seller != User.Identity.Name) return Forbid();
 
         auction.Item.Make = updateAuctionDto.Make ?? auction.Item.Make;
         auction.Item.Model = updateAuctionDto.Model ?? auction.Item.Model;
@@ -98,6 +101,7 @@ public class AuctionsController : ControllerBase
     }
 
     // DeleteAuction
+    [Authorize]
     [HttpDelete("{id}", Name = "DeleteAuction")]
     public async Task<ActionResult> DeleteAuction(Guid id)
     {
@@ -106,7 +110,7 @@ public class AuctionsController : ControllerBase
         .FirstOrDefaultAsync(x => x.Id == id);
         if (auction == null) return NotFound();
 
-        // TODO: Check Seller == Username
+        if (auction.Seller != User.Identity.Name) return Forbid();
 
         _context.Auctions.Remove(auction);
